Tue, Apr 16, 2024, 8:48 PM +08:00

+++ L1/credit.c
@@ -3,6 +3,6 @@
 
 int main(void)
 {
-    unsigned long int number = get_int("Number:");
-    printf("%lu\n", number);
+    unsigned long long int number = get_int("Number:");
+    printf("%llu\n", number);
 }
